requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
attrs==26.1.0
audioop-lts==0.2.2
certifi==2026.2.25
charset-normalizer==3.4.7
discord.py==2.7.1
distro==1.9.0
docstring_parser==0.18.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
jiter==0.14.0
joblib==1.5.3
MLB-StatsAPI==1.9.0
multidict==6.7.1
numpy==2.2.6
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
sniffio==1.3.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
